\name{ABH}
\alias{ABH}
\docType{data}
\title{
The binary outcome trait antibiotic history (ABH) for each subject.
}
\description{
A numeric vector which contains the antibiotic history of each subject. ABH = 0 if the subject did not receive antibiotic treatment in the preceding year. ABH= 1 otherwise. 
}
\usage{data("ABH")}
\format{
  A numeric vector.
}

\references{
HU J, Koh H, He L, Liu M, Blaser J M, Li H(2018). A two-stage microbial association mapping framework with advanced FDR control.
}

\keyword{datasets}
