\name{tax.tab}
\alias{tax.tab}
\docType{data}
\title{
The taxonomic table of the example microbiome data.
}
\description{
A class 'taxonomyTable' from the R package "phyloseq". 

}
\usage{data("tax.tab")}
\format{
A Formal class 'taxonomyTable' [package "phyloseq"] with 1 slot

  ..@ .Data: a matrix where each row represents the taxonomy alignment on Kingdom (Domain), Phylum, Class, Order, Family, Genus and Species represently for each taxon.
}

\keyword{datasets}
