\name{X}
\alias{X}
\docType{data}
\title{
Covariates of the associated microbiome data.
}
\description{
A data frame which contains covariates to be adjusted in the regression models. 
}
\usage{data("X")}
\format{
  A data frame with 100 observations for two covariates.
  \describe{
    \item{\code{age}}{a numeric vector;}
    \item{\code{gender}}{a factor with levels \code{female}, \code{male}.}
  }
}
\references{
HU J, Koh H, He L, Liu M, Blaser J M, Li H(2018). A two-stage microbial association mapping framework with advanced FDR control.
}

\keyword{datasets}
